right before creating brach
at this stage of the project there are two possible filters to be used, the filter bitmap sender and the filter vcam.
bitmap sender works well with splitcam and will be capable of connecting to the web rtc protocole, but it requires splitcam as a driver to be detected. in this brach this is the filter that the dll is registering

the filter vcam is al alternative filter that is currently not being registered, that shows appropiate ultrasound image and is detected as a device that works correctly with teamviewer, however there is some configuration that does not let it work properly with webt rtc and in splitcam and vlc shows a broken blue renderer.